requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-apscheduler==0.6.2
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
